requirements.txt
pynmea2==1.18.0
pyserial==3.5
skyfield==1.42
